Relocated test files to match src file architecture
