+++ tst/core/models/test_yearly_rainfall.py
