+++ tst/core/utils/decorators/test_plots.py
