+++ tst/core/utils/enums/test_seasons.py
